requirements.txt
loguru==0.7.3
boto3==1.37.25
uvicorn==0.34.1
onnxruntime==1.21.0
python-dotenv==1.1.0
python-multipart==0.0.20
fastapi==0.115.12
pipecat-ai==0.0.60
oauth2client==4.1.3
google-genai==1.11.0
google-generativeai==0.8.5
google-cloud==0.34.0
google-cloud-texttospeech==2.26.0
google-cloud-speech==2.32.0
google-cloud-storage==3.1.0
google-cloud-firestore==2.20.1
google-cloud-bigquery==3.31.0